# openapi.yaml
openapi: 3.0.3
info:
  title: Time Tracking API
  description: API for time tracking application with user management, working hours configuration, and time entry tracking
  version: 0.1.0
  contact:
    name: Stefan Remer
    email: stefan.remer@example.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/invite/{token}:
    get:
      tags:
        - Authentication
      summary: Get invite info
      description: Get user info pre-filled from invite token (public)
      operationId: getInviteInfo
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteInfoResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '410':
          description: Invite token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register with invite token
      description: Complete registration using an invite token (public)
      operationId: registerWithInvite
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWithInviteRequest'
      responses:
        '200':
          description: Registration successful, tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '410':
          description: Invite token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users:
    get:
      tags:
        - Users
      summary: List all users
      description: Get list of all users (Admin only)
      operationId: getAllUsers
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Users
      summary: Create new user
      description: Create a new user (Admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Get the profile of the authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Get user details by ID (Admins can get any user, users can only get themselves)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user details (Admins can update any user, users can only update themselves)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete user (Admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/users/{id}/impersonate:
    post:
      tags:
        - Users
      summary: Impersonate a user
      description: Impersonate another user (Admin only, cannot impersonate other admins)
      operationId: impersonateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Impersonation token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Not authorized or cannot impersonate admin users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/users/{id}/invite:
    post:
      tags:
        - Users
      summary: Generate invite link for user
      description: Generate or regenerate an invite/password-reset link for a user (Admin only)
      operationId: generateInviteToken
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Invite link generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Access token expiration time in seconds
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - active
        - state
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        active:
          type: boolean
        state:
          type: string
          enum: [BERLIN, BRANDENBURG]
          description: German state for public holiday calculation
          example: BERLIN
        halfDayHolidaysEnabled:
          type: boolean
          description: Whether half-day holidays (Dec 24 & 31) are enabled for this user
          default: false
          example: false
        pending:
          type: boolean
          description: True if user has no password set yet (awaiting invite completion)
          example: false
        inviteUrl:
          type: string
          description: Invite URL (only present when user is created without password)
          example: http://localhost:5173/register/abc123
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
        - role
      properties:
        email:
          type: string
          format: email
          example: user@timetrack.local
        password:
          type: string
          format: password
          minLength: 6
          example: password123
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Max
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Mustermann
        role:
          type: string
          enum: [ADMIN, USER]
          example: USER
        active:
          type: boolean
          default: true
          example: true
        state:
          type: string
          enum: [BERLIN, BRANDENBURG]
          description: German state for public holiday calculation
          default: BERLIN
          example: BERLIN
        halfDayHolidaysEnabled:
          type: boolean
          description: Whether half-day holidays (Dec 24 & 31) are enabled for this user
          default: false
          example: false
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@timetrack.local
        password:
          type: string
          format: password
          minLength: 6
          description: Only include if changing password
          example: newpassword123
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Max
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Mustermann
        role:
          type: string
          enum: [ADMIN, USER]
          description: Only admins can change roles
          example: USER
        active:
          type: boolean
          description: Only admins can change active status
          example: true
        state:
          type: string
          enum: [BERLIN, BRANDENBURG]
          description: German state for public holiday calculation
          example: BERLIN
        halfDayHolidaysEnabled:
          type: boolean
          description: Whether half-day holidays (Dec 24 & 31) are enabled for this user
          example: false
    InviteTokenResponse:
      type: object
      required:
        - inviteUrl
        - expiresAt
      properties:
        inviteUrl:
          type: string
          description: Full URL for the invite/password-reset link
          example: http://localhost:5173/register/abc123
        expiresAt:
          type: string
          format: date-time
          description: When the invite token expires
    InviteInfoResponse:
      type: object
      required:
        - email
        - firstName
        - lastName
        - isPasswordReset
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        isPasswordReset:
          type: boolean
          description: True if the user already has a password (password reset flow)
    RegisterWithInviteRequest:
      type: object
      required:
        - token
        - password
        - firstName
        - lastName
      properties:
        token:
          type: string
          description: Invite token from the URL
        password:
          type: string
          format: password
          minLength: 6
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message in German
        details:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'